Add inserter

+++ src/lib-header/stdmem.h
@@ -13,7 +13,7 @@
  * 
  * @return Pointer s
 */
-void* memset(void *s, int c, size_t n);
+void* memset(void *s, int c, uint32_t n);
 
 /**
  * C standard memcpy, check man memcpy or
@@ -25,8 +25,8 @@ void* memset(void *s, int c, size_t n);
  * 
  * @return Pointer dest
 */
-void* memcpy(void* restrict dest, const void* restrict src, size_t n);
-void* memcpy_backwards(void* restrict dest, const void* restrict src, size_t n);
+void* memcpy(void* restrict dest, const void* restrict src, uint32_t n);
+void* memcpy_backwards(void* restrict dest, const void* restrict src, uint32_t n);
 
 /**
  * C standard memcmp, check man memcmp or
@@ -38,7 +38,7 @@ void* memcpy_backwards(void* restrict dest, const void* restrict src, size_t n);
  * 
  * @return Integer as error code, zero for equality, non-zero for inequality
 */
-int memcmp(const void *s1, const void *s2, size_t n);
+int memcmp(const void *s1, const void *s2, uint32_t n);
 
 /**
  * C standard memmove, check man memmove or
@@ -50,7 +50,7 @@ int memcmp(const void *s1, const void *s2, size_t n);
  * 
  * @return Pointer dest
 */
-void *memmove(void *dest, const void *src, size_t n);
+void *memmove(void *dest, const void *src, uint32_t n);
 
 //Extras
 uint32_t kmalloc();
